Color update

+++ app/src/main/java/com/example/roleplaybanking/structures/Transaction.java
@@ -8,14 +8,14 @@ import java.util.Date;
 public class Transaction {
     public String sender = "Nobody";
     public String recipient = "Someone";
-    public double amount = 42;
+    public long amount = 42;
     public Timestamp timestamp = new Timestamp(new Date(0));
     public String Notiz;
 
     public Transaction() {
     }
 
-    public Transaction(String sender, String recipient, double amount, Timestamp Time, String Notiz) {
+    public Transaction(String sender, String recipient, long amount, Timestamp Time, String Notiz) {
         this.sender = sender;
         this.recipient = recipient;
         this.amount = amount;
